Fix wrong namespace of extractor classes
